try_lower_option_unwrap_or takes its record declines, scalar match-first attempt and closure class from named helpers
